+++ CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `getQuarter` utility function
+
 ## [2.3.0] - 2026-04-07
 
 ### Added
 
 - `parseEnum` utility function
-- `getCurrentQuarter` utility function
 
 ## [2.2.0] - 2025-10-02